modelo/treinamento_modelo/requirements.txt
# Versões específicas testadas (GTX 1650 + Python 3.8+)
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
numpy==1.24.3
matplotlib==3.7.1
